+++ estampes/parser/gaussian/__init__.py
@@ -10,4 +10,31 @@ glog
     Gaussian output files.
 
 See submodules for details.
+
+The module also provides some conversion functions specific to
+Gaussian, which apply to different types, like the structure of electric
+quadrupole moment
 """
+import typing as tp
+
+
+def g_elquad_LT_to_2D(vec: tp.Sequence[float]) -> tp.List[tp.List[float]]:
+    """Convert electric quadrupole from LT form to 2D tensor.
+
+    Gaussian has a particular way to store internally the electric
+    quadrupole, with the following sequence:
+    XX, YY, ZZ, XY, XZ, YZ.
+    The function builds a list of lists with the correct order.
+
+    Parameters
+    ----------
+    vec
+        Vector of data corresponding to the sequence stored in memory.
+        Note that the list can have a longer size, only the first
+        elements are used.
+    """
+    return [
+        [vec[0], vec[3], vec[4]],
+        [vec[3], vec[1], vec[5]],
+        [vec[4], vec[5], vec[2]]
+    ]