+++ estampes/parser/gaussian/glog/parse_prp.py
@@ -9,6 +9,7 @@ import typing as tp
 from estampes.base import QData, QLabel, \
     ParseKeyError, QuantityError
 from estampes.data.physics import phys_fact
+from estampes.parser.gaussian import g_elquad_LT_to_2D
 
 xyz2id = {'X': 0, 'Y': 1, 'Z': 2}
 
@@ -222,6 +223,142 @@ def parse_1xx_dat(qlab: QLabel, dblock: tp.List[str]) -> QData:
                     raise NotImplementedError()
             else:
                 raise NotImplementedError()
+    elif qlab.label == 102:
+        if isinstance(qlab.rstate, tuple):
+            Si, Sf = qlab.rstate
+            if qlab.derord == 0:
+                dobj.set(unit='hbar.e/me')
+                if Si == 0:
+                    if isinstance(Sf, int):
+                        val = [float(item) for item in dblock[-1][0].split()]
+                    elif Sf == 'a':
+                        val = {}
+                        for i, line in enumerate(dblock[-1]):
+                            val[i+1] = [float(item) for item in line.split()]
+                    else:
+                        raise NotImplementedError(
+                            'Unsupported final state spec.')
+                elif isinstance(Si, int):
+                    if isinstance(Sf, int):
+                        val = [float(item)
+                               for item in dblock[-1][0].split()[2:]]
+                    elif Sf == 'a':
+                        val = {}
+                        for line in dblock[-1]:
+                            cols = line.split()
+                            j, i = int(cols[0]), int(cols[1])
+                            x, y, z = float(cols[2]), float(cols[3]), \
+                                float(cols[4])
+                            val[j] = [x, y, z]
+                    else:
+                        raise NotImplementedError(
+                            'Unsupported final state spec.')
+                elif Si == 'a':
+                    if isinstance(Sf, int):
+                        val = {}
+                        for line in dblock[-1]:
+                            cols = line.split()
+                            j, i = int(cols[0]), int(cols[1])
+                            x, y, z = float(cols[2]), float(cols[3]), \
+                                float(cols[4])
+                            val[i] = [x, y, z]
+                    elif Sf == 'a':
+                        val = {}
+                        for line in dblock[-1]:
+                            cols = line.split()
+                            j, i = int(cols[0]), int(cols[1])
+                            x, y, z = float(cols[2]), float(cols[3]), \
+                                float(cols[4])
+                            if i not in val:
+                                val[i] = {}
+                            val[i][j] = [x, y, z]
+                    else:
+                        raise NotImplementedError(
+                            'Unsupported final state spec.')
+                else:
+                    raise NotImplementedError(
+                        'Unsupported initial state spec.')
+                dobj.set(data=val)
+            else:
+                raise NotImplementedError(
+                    'Electronic transition dipole moments derivatives NYI')
+        else:
+            if qlab.rstate != 'c':
+                # we should check if the state is the right one.
+                raise NotImplementedError(f'Prp {qlab.label} NYI')
+
+            if qlab.derord == 0:
+                raise QuantityError('Magnetic dipole moment not available')
+            else:
+                raise NotImplementedError(f'Prp {qlab.label} NYI')
+    elif qlab.label == 107:
+        if isinstance(qlab.rstate, tuple):
+            Si, Sf = qlab.rstate
+            if qlab.derord == 0:
+                dobj.set(unit='e.a0^2')
+                if Si == 0:
+                    if isinstance(Sf, int):
+                        val = g_elquad_LT_to_2D(
+                            [float(item) for item in dblock[-1][0].split()])
+                    elif Sf == 'a':
+                        val = {}
+                        for i, line in enumerate(dblock[-1]):
+                            val[i+1] = g_elquad_LT_to_2D(
+                                [float(item) for item in line.split()])
+                    else:
+                        raise NotImplementedError(
+                            'Unsupported final state spec.')
+                elif isinstance(Si, int):
+                    if isinstance(Sf, int):
+                        val = g_elquad_LT_to_2D(
+                            [float(item)
+                             for item in dblock[-1][0].split()[2:]])
+                    elif Sf == 'a':
+                        val = {}
+                        for line in dblock[-1]:
+                            cols = line.split()
+                            j, i = int(cols[0]), int(cols[1])
+                            val[j] = g_elquad_LT_to_2D(
+                                [float(item) for item in cols[2:]])
+                    else:
+                        raise NotImplementedError(
+                            'Unsupported final state spec.')
+                elif Si == 'a':
+                    if isinstance(Sf, int):
+                        val = {}
+                        for line in dblock[-1]:
+                            cols = line.split()
+                            j, i = int(cols[0]), int(cols[1])
+                            val[i] = g_elquad_LT_to_2D(
+                                [float(item) for item in cols[2:]])
+                    elif Sf == 'a':
+                        val = {}
+                        for line in dblock[-1]:
+                            cols = line.split()
+                            j, i = int(cols[0]), int(cols[1])
+                            if i not in val:
+                                val[i] = {}
+                            val[i][j] = g_elquad_LT_to_2D(
+                                [float(item) for item in cols[2:]])
+                    else:
+                        raise NotImplementedError(
+                            'Unsupported final state spec.')
+                else:
+                    raise NotImplementedError(
+                        'Unsupported initial state spec.')
+                dobj.set(data=val)
+            else:
+                raise NotImplementedError(
+                    'Electronic transition dipole moments derivatives NYI')
+        else:
+            if qlab.rstate != 'c':
+                # we should check if the state is the right one.
+                raise NotImplementedError(f'Prp {qlab.label} NYI')
+
+            if qlab.derord == 0:
+                raise NotImplementedError(f'Prp {qlab.label} NYI')
+            else:
+                raise NotImplementedError(f'Prp {qlab.label} NYI')
     else:
         raise NotImplementedError(f'Quantity {qlab.label} not yet supported.')